refactor: adapted github.com/bassosimone/clip v0.8.0 code
This diff commits an adaptation of the ./pkg/nflag package of
the github.com/bassosimone/clip v0.8.0 repository.

The changes are breaking compared to the original but the result is
close enough to allow for migration. Main changes are:

1. the default min and max arguments is now zero

2. we have many more value-specific setters

3. the details of how we set values or print help have changes

4. prefixes are now flag specific not parser specific

Yet, the basic usage should be close enough to allow for
mechanical refactoring (except for the min/max defaults that
requires some additional care to be dealt with).